refactor: 페이지 상단 헤더 부분 컴포넌트 화
- 가독성을 위해 컴포넌트 화 진행
