+++ src/App.jsx
@@ -1,17 +1,15 @@
 import { memo } from 'react';
-import { Link, Route, Routes } from 'react-router-dom';
+import { Route, Routes } from 'react-router-dom';
 
 import GlobalStyle from 'GlobalStyle';
 import styled from 'styled-components';
 
-import Button from 'components/common/Button';
 import AddCard from 'pages/AddCard';
 import AddCardComplete from 'pages/AddCardComplete';
-
-import { ReactComponent as Arrow } from 'assets/arrow.svg';
+import Header from 'components/Header';
 
 const Styled = {
-  Page: styled.div`
+  Root: styled.div`
     background: #fff;
     box-sizing: border-box;
     display: flex;
@@ -20,34 +18,19 @@ const Styled = {
     width: 400px;
     height: 757px;
   `,
-  Header: styled.div`
-    align-items: center;
-    display: flex;
-    margin-bottom: 25px;
-  `,
-
-  Title: styled.span`
-    font-size: 16px;
-    margin-left: 18px;
-  `,
 };
 
 function App() {
   return (
     <>
       <GlobalStyle />
-      <Styled.Page>
-        <Styled.Header>
-          <Button size="small">
-            <Arrow />
-          </Button>
-          <Styled.Title>카드 추가</Styled.Title>
-        </Styled.Header>
+      <Styled.Root>
+        <Header title="카드 추가" />
         <Routes>
           <Route path="/card-add" element={<AddCard />} />
           <Route path="/card-add-complete" element={<AddCardComplete />} />
         </Routes>
-      </Styled.Page>
+      </Styled.Root>
     </>
   );
 }
